# project-management-ui/src/app/services/models/appel-offre-request.ts
/* tslint:disable */
/* eslint-disable */
export interface AppelOffreRequest {
  contactInformationRequestDto: string;
  coutRequestDto: number;
  descriptionRequestDto: string;
  endDateRequestDto: string;
  idRequestDto?: number;
  livrablesRequestDto: string;
  projetIdRequestDto: number;
  referenceRequestDto?: string;
  soumissionairesRequestDto: string;
  startDateRequestDto: string;
  statusRequestDto?: string;
  titreRequestDto: string;
  traveauxRequestDto: string;
}