# project-management-ui/src/app/services/models/appel-offre-response.ts
/* tslint:disable */
/* eslint-disable */
export interface AppelOffreResponse {
  contactInformationResponseDto?: string;
  coutResponseDto?: number;
  descriptionResponseDto?: string;
  endDateResponseDto?: string;
  idResponseDto?: number;
  livrablesResponseDto?: string;
  projetIdResponseDto?: number;
  referenceResponseDto?: string;
  soumissionairesResponseDto?: string;
  startDateResponseDto?: string;
  statusResponseDto?: string;
  titreResponseDto?: string;
  traveauxResponseDto?: string;
}