# project-management-ui/src/app/services/models/facture-response.ts
/* tslint:disable */
/* eslint-disable */
export interface FactureResponse {
  contactInformationResponseDto?: string;
  contratIdResponseDto?: number;
  coutResponseDto?: number;
  descriptionResponseDto?: string;
  endDateResponseDto?: string;
  idResponseDto?: number;
  informationFacturationResponseDto?: string;
  referenceResponseDto?: string;
  startDateResponseDto?: string;
  statusResponseDto?: string;
  termePaimentResponseDto?: string;
  titreResponseDto?: string;
}