# project-management-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Ziedzaa",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Ziedzaa",
      "url": "https://some-url.com",
      "email": "ziededucation@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/auth/unlock-account": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "unlockAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/lock-account": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "lockAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/grant-admin": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "grantAdminRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/authenticate": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/appels-offres": {
      "get": {
        "tags": [
          "Appel Offre"
        ],
        "operationId": "findAllAppelOffresByUserStructure",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseAppelOffreResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Appel Offre"
        ],
        "operationId": "saveAppelOffre",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppelOffreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/appels-offres/{appelOffreId}/{status}": {
      "patch": {
        "tags": [
          "Appel Offre"
        ],
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "appelOffreId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/factures/appelOffre/{appelOffreId}": {
      "get": {
        "tags": [
          "Facture"
        ],
        "operationId": "getFacturesByAppelOffreId",
        "parameters": [
          {
            "name": "appelOffreId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFactureResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/contrats/appelOffre/{appelOffreId}": {
      "get": {
        "tags": [
          "Contrat"
        ],
        "operationId": "getContratsByAppelOffreId",
        "parameters": [
          {
            "name": "appelOffreId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseContratResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/activate-account": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "operationId": "confirm",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/appels-offres/{appelOffreId}": {
      "get": {
        "tags": [
          "Appel Offre"
        ],
        "operationId": "findAppelOffreById",
        "parameters": [
          {
            "name": "appelOffreId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppelOffreResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Appel Offre"
        ],
        "operationId": "deleteAppelOffre",
        "parameters": [
          {
            "name": "appelOffreId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/appels-offres/byProjet/{projetId}": {
      "get": {
        "tags": [
          "Appel Offre"
        ],
        "operationId": "findAllAppelOffresByProjetId",
        "parameters": [
          {
            "name": "projetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseAppelOffreResponse"
                }
              }
            }
          }
        }
      }
    },
    "/appels-offres/all": {
      "get": {
        "tags": [
          "Appel Offre"
        ],
        "operationId": "findAllAppelOffres",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseAppelOffreResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FactureResponse": {
        "type": "object",
        "properties": {
          "idResponseDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceResponseDto": {
            "type": "string"
          },
          "titreResponseDto": {
            "type": "string"
          },
          "descriptionResponseDto": {
            "type": "string"
          },
          "coutResponseDto": {
            "type": "number",
            "format": "double"
          },
          "startDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "endDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "statusResponseDto": {
            "type": "string"
          },
          "termePaimentResponseDto": {
            "type": "string"
          },
          "informationFacturationResponseDto": {
            "type": "string"
          },
          "contactInformationResponseDto": {
            "type": "string"
          },
          "contratIdResponseDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AuthenticationRequest": {
        "required": [
          "emailRequestDto",
          "motDePassRequestDto"
        ],
        "type": "object",
        "properties": {
          "emailRequestDto": {
            "type": "string"
          },
          "motDePassRequestDto": {
            "type": "string"
          }
        }
      },
      "RegistrationRequest": {
        "required": [
          "emailRequestDto",
          "motDePasseRequestDto",
          "nomRequestDto",
          "prenomRequestDto"
        ],
        "type": "object",
        "properties": {
          "prenomRequestDto": {
            "pattern": "^[a-zA-ZÀ-ÿ-]{1,15}$",
            "type": "string"
          },
          "nomRequestDto": {
            "pattern": "^[a-zA-ZÀ-ÿ ]{1,20}$",
            "type": "string"
          },
          "emailRequestDto": {
            "type": "string"
          },
          "motDePasseRequestDto": {
            "type": "string"
          }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "tokenResponseDto": {
            "type": "string"
          }
        }
      },
      "AppelOffreRequest": {
        "required": [
          "contactInformationRequestDto",
          "coutRequestDto",
          "descriptionRequestDto",
          "endDateRequestDto",
          "livrablesRequestDto",
          "projetIdRequestDto",
          "soumissionairesRequestDto",
          "startDateRequestDto",
          "titreRequestDto",
          "traveauxRequestDto"
        ],
        "type": "object",
        "properties": {
          "idRequestDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceRequestDto": {
            "type": "string"
          },
          "titreRequestDto": {
            "maxLength": 30,
            "minLength": 2,
            "type": "string"
          },
          "descriptionRequestDto": {
            "type": "string"
          },
          "coutRequestDto": {
            "type": "number",
            "format": "double"
          },
          "startDateRequestDto": {
            "type": "string",
            "format": "date"
          },
          "endDateRequestDto": {
            "type": "string",
            "format": "date"
          },
          "statusRequestDto": {
            "type": "string"
          },
          "traveauxRequestDto": {
            "type": "string"
          },
          "livrablesRequestDto": {
            "type": "string"
          },
          "soumissionairesRequestDto": {
            "type": "string"
          },
          "contactInformationRequestDto": {
            "type": "string"
          },
          "projetIdRequestDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PageResponseFactureResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FactureResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "ContratResponse": {
        "type": "object",
        "properties": {
          "idResponseDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceResponseDto": {
            "type": "string"
          },
          "titreResponseDto": {
            "type": "string"
          },
          "descriptionResponseDto": {
            "type": "string"
          },
          "coutResponseDto": {
            "type": "number",
            "format": "double"
          },
          "startDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "endDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "traveauxResponseDto": {
            "type": "string"
          },
          "livrablesResponseDto": {
            "type": "string"
          },
          "milestonesResponseDto": {
            "type": "string"
          },
          "termPaimentResponseDto": {
            "type": "string"
          },
          "clauseDeResiliasionResponseDto": {
            "type": "string"
          },
          "appelOffreIdResponseDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageResponseContratResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContratResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "AppelOffreResponse": {
        "type": "object",
        "properties": {
          "idResponseDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceResponseDto": {
            "type": "string"
          },
          "titreResponseDto": {
            "type": "string"
          },
          "descriptionResponseDto": {
            "type": "string"
          },
          "coutResponseDto": {
            "type": "number",
            "format": "double"
          },
          "startDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "endDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "statusResponseDto": {
            "type": "string"
          },
          "traveauxResponseDto": {
            "type": "string"
          },
          "livrablesResponseDto": {
            "type": "string"
          },
          "soumissionairesResponseDto": {
            "type": "string"
          },
          "contactInformationResponseDto": {
            "type": "string"
          },
          "projetIdResponseDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageResponseAppelOffreResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppelOffreResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}